getting of dev and working judge page
