+++ modules/hackathons/client/controllers/list-projects.client.controller.js
@@ -1,0 +1,17 @@
+(function () {
+    'use strict';
+  
+    angular
+      .module('hackathons')
+      .controller('ProjectsListController', ProjectsListController);
+  
+    ProjectsListController.$inject = ['ProjectsService'];
+  
+    function ProjectsListController(ProjectsService) {
+      var vm = this;
+  
+      vm.hackathons = ProjectsService.query();
+     
+    }
+  }());
+
