--- modules/projects/client/controllers/list-projects.client.controller.js
@@ -1,15 +1,0 @@
-(function () {
-  'use strict';
-
-  angular
-    .module('projects')
-    .controller('ProjectsListController', ProjectsListController);
-
-  ProjectsListController.$inject = ['ProjectsService'];
-
-  function ProjectsListController(ProjectsService) {
-    var vm = this;
-
-    vm.projects = ProjectsService.query();
-  }
-}());
