--- modules/projects/client/projects.client.module.js
@@ -1,5 +1,0 @@
-(function (app) {
-  'use strict';
-
-  app.registerModule('projects');
-}(ApplicationConfiguration));
