--- modules/projects/client/views/form-project.client.view.html
@@ -1,24 +1,0 @@
-<section>
-  <div class="page-header">
-    <h1>{{vm.project._id ? 'Edit Project' : 'New Project'}}</h1>
-  </div>
-  <div class="col-md-12">
-    <form name="vm.form.projectForm" class="form-horizontal" ng-submit="vm.save(vm.form.projectForm.$valid)" novalidate>
-      <fieldset>
-        <div class="form-group" show-errors>
-          <label class="control-label" for="name">Name</label>
-          <input name="name" type="text" ng-model="vm.project.name" id="name" class="form-control" placeholder="Name" required>
-          <div ng-messages="vm.form.projectForm.name.$error" role="alert">
-            <p class="help-block error-text" ng-message="required">Project name is required.</p>
-          </div>
-        </div>
-        <div class="form-group">
-          <button type="submit" class="btn btn-default">{{vm.project._id ? 'Update' : 'Create'}}</button>
-        </div>
-        <div ng-show="vm.error" class="text-danger">
-          <strong ng-bind="vm.error"></strong>
-        </div>
-      </fieldset>
-    </form>
-  </div>
-</section>
