--- modules/projects/server/config/projects.server.config.js
@@ -1,14 +1,0 @@
-'use strict';
-
-/**
- * Module dependencies
- */
-var path = require('path'),
-  config = require(path.resolve('./config/config'));
-
-/**
- * Projects module init function.
- */
-module.exports = function (app, db) {
-
-};
