--- modules/projects/tests/client/list-projects.client.controller.tests.js
@@ -1,91 +1,0 @@
-(function () {
-  'use strict';
-
-  describe('Projects List Controller Tests', function () {
-    // Initialize global variables
-    var ProjectsListController,
-      $scope,
-      $httpBackend,
-      $state,
-      Authentication,
-      ProjectsService,
-      mockProject;
-
-    // The $resource service augments the response object with methods for updating and deleting the resource.
-    // If we were to use the standard toEqual matcher, our tests would fail because the test values would not match
-    // the responses exactly. To solve the problem, we define a new toEqualData Jasmine matcher.
-    // When the toEqualData matcher compares two objects, it takes only object properties into
-    // account and ignores methods.
-    beforeEach(function () {
-      jasmine.addMatchers({
-        toEqualData: function (util, customEqualityTesters) {
-          return {
-            compare: function (actual, expected) {
-              return {
-                pass: angular.equals(actual, expected)
-              };
-            }
-          };
-        }
-      });
-    });
-
-    // Then we can start by loading the main application module
-    beforeEach(module(ApplicationConfiguration.applicationModuleName));
-
-    // The injector ignores leading and trailing underscores here (i.e. _$httpBackend_).
-    // This allows us to inject a service but then attach it to a variable
-    // with the same name as the service.
-    beforeEach(inject(function ($controller, $rootScope, _$state_, _$httpBackend_, _Authentication_, _ProjectsService_) {
-      // Set a new global scope
-      $scope = $rootScope.$new();
-
-      // Point global variables to injected services
-      $httpBackend = _$httpBackend_;
-      $state = _$state_;
-      Authentication = _Authentication_;
-      ProjectsService = _ProjectsService_;
-
-      // create mock article
-      mockProject = new ProjectsService({
-        _id: '525a8422f6d0f87f0e407a33',
-        name: 'Project Name'
-      });
-
-      // Mock logged in user
-      Authentication.user = {
-        roles: ['user']
-      };
-
-      // Initialize the Projects List controller.
-      ProjectsListController = $controller('ProjectsListController as vm', {
-        $scope: $scope
-      });
-
-      // Spy on state go
-      spyOn($state, 'go');
-    }));
-
-    describe('Instantiate', function () {
-      var mockProjectList;
-
-      beforeEach(function () {
-        mockProjectList = [mockProject, mockProject];
-      });
-
-      it('should send a GET request and return all Projects', inject(function (ProjectsService) {
-        // Set POST response
-        $httpBackend.expectGET('api/projects').respond(mockProjectList);
-
-
-        $httpBackend.flush();
-
-        // Test form inputs are reset
-        expect($scope.vm.projects.length).toEqual(2);
-        expect($scope.vm.projects[0]).toEqual(mockProject);
-        expect($scope.vm.projects[1]).toEqual(mockProject);
-
-      }));
-    });
-  });
-}());
