--- modules/projects/tests/e2e/projects.e2e.tests.js
@@ -1,10 +1,0 @@
-'use strict';
-
-describe('Projects E2E Tests:', function () {
-  describe('Test Projects page', function () {
-    it('Should report missing credentials', function () {
-      browser.get('http://localhost:3001/projects');
-      expect(element.all(by.repeater('project in projects')).count()).toEqual(0);
-    });
-  });
-});
