--- modules/projects/tests/server/project.server.model.tests.js
@@ -1,67 +1,0 @@
-'use strict';
-
-/**
- * Module dependencies.
- */
-var should = require('should'),
-  mongoose = require('mongoose'),
-  User = mongoose.model('User'),
-  Project = mongoose.model('Project');
-
-/**
- * Globals
- */
-var user,
-  project;
-
-/**
- * Unit tests
- */
-describe('Project Model Unit Tests:', function() {
-  beforeEach(function(done) {
-    user = new User({
-      firstName: 'Full',
-      lastName: 'Name',
-      displayName: 'Full Name',
-      email: 'test@test.com',
-      username: 'username',
-      password: 'password'
-    });
-
-    user.save(function() {
-      project = new Project({
-        name: 'Project Name',
-        user: user
-      });
-
-      done();
-    });
-  });
-
-  describe('Method Save', function() {
-    it('should be able to save without problems', function(done) {
-      this.timeout(0);
-      return project.save(function(err) {
-        should.not.exist(err);
-        done();
-      });
-    });
-
-    it('should be able to show an error when try to save without name', function(done) {
-      project.name = '';
-
-      return project.save(function(err) {
-        should.exist(err);
-        done();
-      });
-    });
-  });
-
-  afterEach(function(done) {
-    Project.remove().exec(function() {
-      User.remove().exec(function() {
-        done();
-      });
-    });
-  });
-});
